backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
pydantic-settings
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
